@@ -1,4 +1,4 @@
-export interface Error {
+export interface AppError {
   error: string;
   message: string;
   code: number;
