@@ -1,9 +1,51 @@
 import {Component, EventEmitter, Input, Output} from '@angular/core';
-import {MatTableDataSource} from '@angular/material/table';
+import {
+  MatCell,
+  MatCellDef,
+  MatColumnDef,
+  MatHeaderCell,
+  MatHeaderCellDef,
+  MatHeaderRow,
+  MatHeaderRowDef,
+  MatRow,
+  MatRowDef,
+  MatTable,
+  MatTableDataSource
+} from '@angular/material/table';
 import {Observable} from 'rxjs';
+import {MatCard, MatCardContent, MatCardTitle} from '@angular/material/card';
+import {NgForOf, NgIf} from '@angular/common';
+import {UppercaseWords} from '@shared/pipes/UppercaseWordsPipe';
+import {MatIconButton} from '@angular/material/button';
+import {MatSuffix} from '@angular/material/form-field';
+import {MatIcon} from '@angular/material/icon';
+import {MatSort} from '@angular/material/sort';
 
 @Component({
   selector: 'app-crud',
+  standalone:true,
+  imports: [
+    MatCard,
+    MatCardTitle,
+    MatCardContent,
+    MatTable,
+    NgIf,
+    MatHeaderRow,
+    MatHeaderRowDef,
+    MatRowDef,
+    MatRow,
+    MatColumnDef,
+    MatHeaderCell,
+    UppercaseWords,
+    MatCell,
+    MatCellDef,
+    MatHeaderCellDef,
+    NgForOf,
+    MatIconButton,
+    MatSuffix,
+    MatIcon,
+    MatSort
+  ],
   templateUrl: 'crud.component.html'
 })
 export class CrudComponent {
@@ -54,7 +96,8 @@ export class CrudComponent {
   onDelete(item): void {
     this.delete.emit(item);
   }
-  isArray(obj : any ) {
+
+  isArray(obj: any) {
     return Array.isArray(obj)
   }
 
