@@ -1,16 +1,41 @@
-import {Component, EventEmitter, Input, Output} from '@angular/core';
+import {booleanAttribute, Component, EventEmitter, Input, Output} from '@angular/core';
 import {of} from 'rxjs';
+import {MatFormField, MatLabel, MatSuffix} from '@angular/material/form-field';
+import {MatIcon} from '@angular/material/icon';
+import {MatAutocomplete, MatAutocompleteTrigger, MatOption} from '@angular/material/autocomplete';
+import {AsyncPipe, NgFor, NgIf} from '@angular/common';
+import {MatIconButton} from '@angular/material/button';
+import {MatTooltip} from '@angular/material/tooltip';
+import {FormsModule} from '@angular/forms';
+import {MatInput} from '@angular/material/input';
 
 @Component({
   selector: 'app-search',
+  standalone:true,
   templateUrl: 'search.component.html',
+  imports: [
+    MatFormField,
+    MatIcon,
+    MatAutocomplete,
+    MatOption,
+    AsyncPipe,
+    MatIconButton,
+    MatLabel,
+    MatTooltip,
+    FormsModule,
+    MatAutocompleteTrigger,
+    MatInput,
+    MatSuffix,
+    NgIf,
+    NgFor
+  ],
   styleUrls: ['search.component.css']
 })
 export class SearchComponent {
   @Input() title = 'Search';
   @Input() key: string;
   @Input() keys = of([]);
-  @Input() obligatory = "false";
+  @Input({transform: booleanAttribute}) obligatory = false;
 
   @Output() keyChange = new EventEmitter<string>();
   @Output() renew = new EventEmitter<any>();
@@ -31,5 +56,9 @@ export class SearchComponent {
 
   public onChange(value: string): void {
     this.keyChange.emit(value);
+  }
+
+  trackOption(index: number, option: any): any {
+    return option;
   }
 }
