@@ -2,17 +2,38 @@ import {Component} from '@angular/core';
 import {Router} from '@angular/router';
 
 import {AuthService} from '@core/auth.service';
-import {MatDialog} from '@angular/material/dialog';
+import {MatDialog, MatDialogActions, MatDialogClose, MatDialogContent, MatDialogTitle} from '@angular/material/dialog';
+import {FormsModule} from '@angular/forms';
+import {MatFormField, MatSuffix} from '@angular/material/form-field';
+import {MatInput} from '@angular/material/input';
+import {MatButton, MatIconButton} from '@angular/material/button';
+import {MatIcon} from '@angular/material/icon';
+import {NgIf} from '@angular/common';
 
 @Component({
+  standalone:true,
   templateUrl: 'login-dialog.component.html',
+  imports: [
+    MatDialogTitle,
+    MatDialogContent,
+    FormsModule,
+    MatSuffix,
+    MatInput,
+    MatButton,
+    MatIconButton,
+    MatFormField,
+    MatIcon,
+    MatDialogActions,
+    MatDialogClose,
+    NgIf
+  ],
   styleUrls: ['./dialog.component.css']
 })
 export class LoginDialogComponent {
   mobile: number;
   password: string;
 
-  constructor(private auth: AuthService, private router: Router, private dialog: MatDialog) {
+  constructor(private readonly auth: AuthService, private readonly router: Router, private readonly dialog: MatDialog) {
   }
 
   login(): void {
