@@ -1,10 +1,10 @@
 import {environment} from '@env';
 
 export class EndPoints {
-  static PROVIDERS = environment.REST_CORE + '/providers';
-  static ARTICLES = environment.REST_CORE + '/articles';
-  static CASHIERS = environment.REST_CORE + '/cashiers';
-  static CASHIERS_LAST = EndPoints.CASHIERS + '/last';
-  static TICKETS = environment.REST_CORE + '/tickets';
-  static COMPLAINTS = environment.REST_CUSTOMER_SUPPORT + '/complaints';
+  static readonly PROVIDERS = environment.REST_CORE + '/providers';
+  static readonly ARTICLES = environment.REST_CORE + '/articles';
+  static readonly CASHIERS = environment.REST_CORE + '/cashiers';
+  static readonly CASHIERS_LAST = EndPoints.CASHIERS + '/last';
+  static readonly TICKETS = environment.REST_CORE + '/tickets';
+  static readonly COMPLAINTS = environment.REST_CUSTOMER_SUPPORT + '/complaints';
 }
