@@ -1,9 +1,14 @@
 import {Observable, of} from 'rxjs';
 import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {SharedProviderService} from './services/shared.provider.service';
+import {SearchComponent} from '@shared/components/search.component';
 
 @Component({
   selector: 'app-search-by-company',
+  standalone:true,
+  imports: [
+    SearchComponent
+  ],
   templateUrl: './search-by-company.component.html'
 })
 export class SearchByCompanyComponent {
@@ -12,7 +17,7 @@ export class SearchByCompanyComponent {
   @Input() company: string;
   @Output() companyChange = new EventEmitter<string>();
 
-  constructor(private sharedProviderService: SharedProviderService) {
+  constructor(private readonly sharedProviderService: SharedProviderService) {
   }
 
   public onSelect(): void {
